[State Sync] Rename helper.rs file to mock_executor_proxy.rs
