+++ state-synchronizer/src/tests/mod.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(test)]
-mod helpers;
-#[cfg(test)]
 mod integration_tests;
+#[cfg(test)]
+mod mock_executor_proxy;
 #[cfg(test)]
 mod mock_storage;
 #[cfg(test)]
